+++ src/components/ui/BaseCardCitySkeleton.vue
