+++ src/components/ui/BaseCardGrid.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="w-full max-w-screen-2xl m-auto grid grid-cols-1 auto-rows-cards gap-4 mt-20 lg:grid-cols-2 lg:gap-8 xl:grid-cols-3 xl:gap-8">
-    <CardCity
+    <BaseCardCity
       v-for="city in weatherStore.getCities"
       :key="city.id"
       :city="city"
@@ -8,14 +8,14 @@
       :is-loading="commonStore.getStatus === 'updating'"
     />
 
-    <CardCitySkeleton v-if="commonStore.getStatus === 'loading'" />
+    <BaseCardCitySkeleton v-if="commonStore.getStatus === 'loading'" />
   </div>
 </template>
 
 <script setup lang="ts">
 import { useCommonStore, useWeatherStore } from '@/stores';
-import CardCity from '@/components/ui/CardCity.vue';
-import CardCitySkeleton from '@/components/ui/CardCitySkeleton.vue';
+import BaseCardCity from '@/components/ui/BaseCardCity.vue';
+import BaseCardCitySkeleton from '@/components/ui/BaseCardCitySkeleton.vue';
 
 const commonStore = useCommonStore();
 const weatherStore = useWeatherStore();
