+++ src/components/ui/BaseSearchInput.vue
