+++ src/components/ui/BaseSearchSuggestionItem.vue
